Fixed typo in README.md
Adds an 's' to method on line 102 where it should be plural, change adds grammatical clarity and improves flow for the reader. Nothing else is changed:

OLD LINE: The difference for the immutable collections is that method which would mutate
the collection, like `push`, `set`, `unshift` or `splice` instead return a new
immutable collection.

NEW LINE: The difference for the immutable collections is that methods which would mutate
the collection, like `push`, `set`, `unshift` or `splice` instead return a new
immutable collection.