Fix fromEntries return type
Fixes #31.

Previously, grouping a keyed sequence would return each group as an IndexedSequence with nonnumeric keys, which was wrong. This was caused by fromEntries returning an IndexedSequence instead of a keyed sequence. With this, fromEntries always returns a keyed sequence (which matches the type definition) and Sequence's groupBy returns the correct type. (Calling groupBy on an IndexedSequence still gives each group as an IndexedSequence, as before.)

Test Plan: jest