Cursors!
Thanks for the suggestion, #41!

This is designed to be a minimum viable implementation of Cursors - and since I just wrote it, it is not in production anywhere yet. Use with caution, caveat emptor and all that. Also, it means the API is subject to change as feedback is provided (provide feedback!)