Fix grammar and punctuation in README.md
I enjoy using immutable a lot and wanted to contribute.

I found the expression in Line 333-335 of the README.md slightly unclear, and made the following changes:

BEFORE:
There is a performance penalty paid every time you create a new immutable object via applying a mutation. If you need to apply a series of mutations immutable gives you the ability to create a temporary mutable copy of a collection and applying a batch of mutations in a highly performant

AFTER:
Applying a mutation to create a new immutable object will result in a performance penalty.
If you need to apply a series of mutations, `immutable` gives you the ability to create a
temporary mutable copy of a collection and apply a batch of mutations

Nothing else was changed.