Add steps to CONTIBUTING.md
These were the stumbling blocks that were a bit annoying to get going
with, having never used grunt before. I was also ensure as to whether or
not I needed to commit the built JS/interface; inspection of previous
commits indicated that was expected.

I'd understand if this may be superfluous, but if not, merge away!