Strict Iterable/Seq constructors.
This makes the Seq constructors (and by proxy, the Collection constructors) stricter. List and Set no longer accept Objects with discarded keys, as that is of dubious value and more often could represent a bug or programmer error. When a KeyedSeq is provided to List and Set, it should behave as an iterable and provide [K,V] tuples rather than discarding keys. This makes between-version operation smoother and makes Immutable's Iterator behave more similarly to generic Iterators.

The toXXX() methods have preserved their behavior.

The behavior of "array-like" values is stricter as well. Only Indexed (List) and Set iterables will consume any array-like as an array (such as `arguments` or String). Keyed iterables will always favor Object keys, which removes a pitfall where trying to construct a Map where "length" is a key would fail.

Finally, this improves `concat` to follow these rules based on the parent iterable type. Previously concatting an object to a List would cause the keys of the object to be treated as a list, which in the context of `concat`'s API of `concat(...valueOrArrays)` made concatting a single value dangerous affair. `concat` now expressly accepts `Iterables`, `Array` or singleton "value", and will not attempt to coerce any value other than Array into an Iterable.