Replace Smash with ES6 modules
This is a pretty huge diff, and deserves some explanation.

 * Replaces the smash-specific "import" declarations and the jslint @globals and @exports docblocks with ES6-module-style "import" and "export" declarations.

 * Excludes `smash` from the dev dependencies, includes `esperanto` - a bundler similar to `smash` but which operates on ES6-modules.

 * Broke the implicit circular dependencies by moving most of Iterable's implementation to IterableImpl.js, responsible for building the prototype of Iterable.

 * There were a few other implicit circular dependencies solved by moving shared utility functions into their own modules.

This is also the first step in this direction. While the bundle builds and all tests pass, this has introduced awkwardness and a few (negligable) additional bytes to the bundle. There is also ample opportunity for further refactoring of the larger modules to better organize and make circular dependencies more clear.

Closes #273