Use uncommon key for canDefineProperty test
Some versions of Chrome have a bug where in some conditions, setting a property as non-enumerable can result in that property being non enumerable for all objects, as reported by #295. As a work-around, we should be using a property name that is far less likely to be used as an enumerable property in the future.