Add specialized Map#map to keep ref. equality
The point is to keep the referential equality when no value changed. This let
one call `Map#map` without afterthoughts about optimization. For example, let's
imagine that we trigger an interface render when a `Map` changed; we want to
avoid the cost of re-rendering when no value changed. Without referential
equality, we'd have a write some kind of custom algorithm to compare the values
returned by `mapper`.

The performance of this function is no worse, as the generalized
`IterableImpl#map` always reifies the mapped sequence anyway. On the other hand,
the fact we're applying modifications to the existing `Map` mean we're
benefiting from the internal structure optimizations.

As a side note, I find it interesting to see how a custom `map` function looks
significantly simpler than going through the generalized operation on `Seq`.
Granted, we lose the unification of sequence transformation algorithms, but
the original problem this changeset addresses proves it's hardly possible to
totally abstract a `Map` as a `Seq`: a `Seq` loses the concept of concreteness.

A similar fix can be used for `filter`, at least.

See #285.