Ensure values of non ES5 types can be hashed
Fixes #579

ES6 introduces both 'symbol' as the possible result of a `typeof` expression, as well as makes room for the possibility of future different values from non-standard exotics or presumably from ES7 proposed additions.

http://www.ecma-international.org/ecma-262/6.0/#sec-typeof-operator

This amends the `hash` function to only perform hashJSObj if the value is in fact a typeof 'object'. Otherwise as a fallback, we rely on the existence of a 'toString' function, and the value is hashed as if it were a string using the return of toString. If toString has not been defined, then an error is raised.