Update dependencies
This updated the majority of dev dependencies to their latest versions, excepting any which cause significant issue (acorn / estraverse) and those related to linting and testing which will carry larger updates soon.

In addition, those dependencies changed have moved to declaring an exact version rather than a version range. This improves the assurance that a development environment which works at one point will continue to work in the future.