Replace custom equality tester for custom matcher
Cursor test was having issues with using a custom equality tester that were sporadically failing to apply. This replaces with a slightly more explicit custom matcher, which is probably preferable anyhow as it more explicitly declares what is being tested.

Finally, this replaces the `lastCalledWith` matcher with more explicit matchers for each argument. This lets us apply different assertions to each argument.