Use polymorphic 'this' return types for Map methods
## The Problem

As a simplified example, consider the following `ABClass`, which extends from `ABRecord` (itself instantiated with `Record({ a: 1, b: 2 })`):

```ts
const ABRecord = Record({ a: 1, b: 2 });

class ABClass extends ABRecord {
  a: number;
  b: number;

  setA(a: number) {
    return this.set('a', a);
  }

  setB(b: number) {
    return this.set('b', b);
  }
}
```

Because the return type of `Map.set` is specified explicitly as `Map<K, V>`, the following attempt at chaining the method calls will alert the TS compiler:

```ts
const ab = new ABClass({ b: 3 });
const ab2 = ab.setA(5).setB(6);
console.log(ab2.a, ab2.b);
```

As a result, we must cast the return types of `ABClass.setA` and `ABClass.setB` methods explicitly to `ABClass` in order to satisfy the TS compiler.

---

The situation is more complicated when calling `withMutations`:

```ts
const ab3 = ab.withMutations(
  mutable => mutable.set('a', mutable.a + 1).set('b', mutable.b + 1));
```

The above example fails to compile because `mutable` is typed to `Map<string, any>`, which does not have the properties `a` and `b`, so we have to either access them with `get`, or cast mutable to `ABClass`, as well as having to cast the return type that will be assigned to `ab3`.

## The Solution

This PR addresses those problems, by using polymorphic `this` return types for `Map.*` methods that result in a new `Map` of the same type.