Make header nav links dark gray; increase contrast for usability
I noticed that the contrast between the navigation links and the
background on the Immutable.js landing page is very subtle - by
using higher contrast we can increase usability and accessibility.

There is a handy tool for testing contrast for accessibility;
http://webaim.org/resources/contrastchecker/

On the landing page the links are white on a light gray background.
We initially set the link color to match the main font color on the
page, which is #626466. However, this was not quite enough contrast.
Bumping the font color a step darker gave us #525455, which should be
very close but have a high enough contrast to be easily readable.

The dark gray (#525455) on the light blue background was also not high
enough contrast, so we lightened the light blue background.

We considered leaving the links white on the blue background, and only
changing them on the landing page where the background is light gray.
However, to get enough contrast with a white font the light blue
background would have to be darkened beyond recognition.

Screenshots will be attached in the description or comments.

I'm open to adjusting this as needed, and thanks for your review!

More ideas:
- We could make the main copy font color match the nav links
- We could use a dark blue instead of dark gray for the nav links