Fix OrderedSet and OrderedMap flow annotations
Unfortunately extending Set and Map caused issues with OrderedSet
and OrderedMap flow annotations. For example,

```
const orderedSet: OrderedSet<number> = OrderedSet.of(1,2,3);
```

was not valid, because `of` was incorrectly assumed to return a
`Set<number>` rather than an `OrderedSet<number>`.

The fix was to copy and paste the Set and Map flow annotations,
adjusting for OrderedSet and OrderedMap as appropriate.

See #1015 for more details.