Change toJSON to be a shallow operation.
The primary use for toJSON is to be called from JSON.stringify to convert exotics into JSON-ready data. Since JSON.stringify continues to recursively operate on the return value of toJSON, then it shouldn't deeply convert.

This is technically a breaking change since previously toJSON was advertised as an alias for toJS. The behavior of toJS doesn't change.

This results in a minor byte-size simplification for the library and a minor speed boost for JSON.stringify over an immutable collection.