Add back deleteIn() and removeIn() to Record (#1179)
As part of the RFC to refactor Records [1], the methods `deleteIn` and
`removeIn` were removed but the Typescript and Flow definitions were
still present. This PR will bring back these methods to deeply
manipulate a tree of immutable records. This is especially useful when
you build a larger state out of immutable components and want to remove
a list entry deeply within.

[1]: https://github.com/facebook/immutable-js/commit/768151abf9e360b4d7a7e6e93e03470cd29ff5e9