Loosen type of Keyed.flatMap and descendants
`Keyed.flatMap` says that `mapper` is allowed to change the key type and the
value type, but `Collection.keyMap` says that `mapper` is only allowed to
change the value type. This commit resolves the typing problem by making
the return type's type arguments both `any`. It does not resolve the
mismatch in the implementation of Map.flatMap, for example.