Set types:K type parameter is now never
The previous type, Set<T> extends Collection<T, T>, results in an
incorrect type for `map` because `map` is not allowed to change the key
type, only the value type. But map's previous return type of Set<T>
asserted that both the key *and* the value type changed. Now the key
type is always `never`.