Stop the iteration of an exhausted, unknown-sized sequence when slicing
This PR fixes a bug with the slice operator that caused non value
iterators of an unknown sized sequence to continue although the source
sequence was exhausted, eventually accessing the value of a key/value
pair that does not exist.

The issue did not happen with fixed sized sequences since we can
calculate resulting size in advance (as `sliceSize`) and bail out early.
For an unknown sized sequence however, we can not know this size in
advance (see this comment: [1]).

To solve this, we now bail out when we encounter a completed iterator
before proceeding to process it further.

[1]: https://github.com/facebook/immutable-js/blob/27b231c8952366ddb96ab2a20d1eb351b3c8719e/src/Operations.js#L427-L428