BREAKING: concat Lists when merging deeply
Fixes a long standing issue where "merge" did not treat List as an expected monoid form, overwriting entries as if it were a Map instead. This *changes* `merge` on List to be an alias of `concat` (much like `merge` is an alias for `union` on Set). Other functions like `mergeDeep` and mergeWith` have been *removed* from List, since they no longer make sense with this behavior. When `mergeDeep` is called on a Map, nested Lists are now concatenated instead of overwritten.

Fixes #406