BREAKING: Create new List instead of new Map in updateIn keypath.
If any keys in `keyPath` do not exist, a new Map is created. This is often confusing when using numeric keys - which typically imply using a List instead of a Map.

This now looks at the key, if it is a number then a List is created instead of a Map.

This is breaking since previously this would have created a Map instead of a List. Existing code relying on creating `Map<number, *>` should convert code to explicitly create missing Maps before calling `updateIn()`.