BREAKING: Use === for deep merge equality.
This rectifies an inconsistent behavior between `x.merge(y)` and `x.mergeDeep(y)` where merge would use === on leaf values to determine return-self optimizations, while mergeDeep would use is(). Not only was this inconsistent, it was the only place in the library this behavior existed - where everwhere else `===` is used to determine return-self optimization.

Also minor refactor to save a handful of gz-bytes and reduce overhead.