Use prettier for tests
When working on projects that use prettier, I enable the format-on-save
option in my editor of choice which run prettier whenever I save a file.

This always updated the formatting of the tests files when working on
Immutable.js, since those files were not formatted using prettier yet.
Prettier added [typescript support][1] in the mean time, so I thought
it was a good idea to keep the whole code base consistent and to allow
everyone to use format-on-save without getting annoyed.

Since there are not a lot of PRs open, I think it's a good time to merge
this.

[1]: https://github.com/prettier/prettier/releases/tag/1.4.0