Simplify typescript definition files
This removes the build step which converts an "ambient" type definition file into a "non-ambient" one. Since our local tests all run against the source file, this ensures the tested file is exactly the same as the exported one, and updates the exports to support all "UMD" use cases, both ambient (sourced file with a `/// <reference>` and non-ambient module exports.

As a result, we no longer need to explicity mark each type within the exported module

I'm ~90% sure this is the correct way to set up the module and UMD exports, but would be very happy to have a second set of eyes on this.

This also changes the upper-case source files to lower-case so they can be directly copied during build.

Tested by creating a new vanilla typescript project and importing the built npm directory and tested a bunch of imports of Immutable to ensure the types were being correctly referenced.