reverse empty argument check in constructors
The most common way to call a constructor to make an empty collection is
to call it without arguments, then value is undefined.

With this commit the first check to see if the value passed to the
constructor is not set is for undefined, the most common way to call
it.

This way the condition will only have to evaluate the left side for the
common case.