Revert previous assertion as it introduced a regression
It seems reasonable to assert that keys shouldn't be undefined,
since keys are supposed to be of type PropertyKey (string | number | symbol).
But in practice, JavaScript allows undefined as a key by implicitly
converting it to the string "undefined".

This caused a regression specifically in the `groupBy` function, where the
grouper callback can return undefined for some items. That behavior was
previously supported and is now broken by the assertion.

Long-term, it probably makes sense to tighten this up and only allow
real PropertyKeys, but that’ll need a more careful cleanup.